Moved List into lists
