+++ lib/list/Place.ts
@@ -1,28 +1,22 @@
-import { List } from "../List";
-import { AgeList } from "../list/age/age";
-import { AlignmentList } from "../list/alignment/alignment";
-import { ArmorList } from "../list/armor/armor";
-import {
-    BackgroundList,
-    BackgroundProperties,
-} from "../list/background/background";
-import { CharacteristicList } from "../list/characteristics/characteristics";
-import { ClassList, ClassProperties } from "../list/class/class";
-import { EyeList } from "../list/eye/eye";
-import { HairList } from "../list/hair/hair";
-import { LanguageList } from "../list/languages/languages";
-import { MaturityList } from "../list/maturity/maturity";
-import { MotivationList } from "../list/motivation/motivation";
-import { NobilityList } from "../list/nobility/nobility";
-import {
-    ProfessionList,
-    ProfessionProperties,
-} from "../list/profession/profession";
-import { RaceList, RaceProperties } from "../list/race/race";
-import { SexList } from "../list/sex/sex";
-import { SkinList } from "../list/skin/skin";
-import { TitleList } from "../list/title/title";
-import { WeaponList } from "../list/weapon/weapon";
+import { List } from "./List";
+import { AgeList } from "./age/age";
+import { AlignmentList } from "./alignment/alignment";
+import { ArmorList } from "./armor/armor";
+import { BackgroundList, BackgroundProperties } from "./background/background";
+import { CharacteristicList } from "./characteristics/characteristics";
+import { ClassList, ClassProperties } from "./class/class";
+import { EyeList } from "./eye/eye";
+import { HairList } from "./hair/hair";
+import { LanguageList } from "./languages/languages";
+import { MaturityList } from "./maturity/maturity";
+import { MotivationList } from "./motivation/motivation";
+import { NobilityList } from "./nobility/nobility";
+import { ProfessionList, ProfessionProperties } from "./profession/profession";
+import { RaceList, RaceProperties } from "./race/race";
+import { SexList } from "./sex/sex";
+import { SkinList } from "./skin/skin";
+import { TitleList } from "./title/title";
+import { WeaponList } from "./weapon/weapon";
 
 interface Lists {
     maturity: MaturityList | List;
@@ -68,7 +62,7 @@ export class Place {
                 title: new TitleList(),
                 weapon: new WeaponList(),
                 background: new BackgroundList(),
-                languages: new LanguageList(),
+                languages: new LanguageList()
             },
             lists || {}
         );
