+++ lib/list/alignment/alignment.ts
@@ -1,4 +1,4 @@
-import { List } from "../../List";
+import { List } from "../List";
 
 export class AlignmentList extends List {
     constructor() {
