+++ lib/list/armor/armor.ts
@@ -1,4 +1,4 @@
-import { List } from "../../List";
+import { List } from "../List";
 
 export class ArmorList extends List {
     constructor() {
