+++ lib/list/characteristics/characteristics.ts
@@ -1,4 +1,4 @@
-import { List } from "../../List";
+import { List } from "../List";
 
 export class CharacteristicList extends List {
     constructor() {
