+++ lib/list/eye/eye.ts
@@ -1,4 +1,4 @@
-import { List } from "../../List";
+import { List } from "../List";
 
 export class EyeList extends List {
     constructor() {
