+++ lib/list/hair/hair.ts
@@ -1,4 +1,4 @@
-import { List } from "../../List";
+import { List } from "../List";
 
 export class HairList extends List {
     constructor() {
