+++ lib/list/languages/languages.ts
@@ -1,4 +1,4 @@
-import { List } from "../../List";
+import { List } from "../List";
 
 export class LanguageList extends List {
     constructor() {
