+++ lib/list/motivation/motivation.ts
@@ -1,4 +1,4 @@
-import { List } from "../../List";
+import { List } from "../List";
 
 export class MotivationList extends List {
     constructor() {
