+++ lib/list/nobility/nobility.ts
@@ -1,4 +1,4 @@
-import { List } from "../../List";
+import { List } from "../List";
 
 export class NobilityList extends List {
     constructor() {
