+++ lib/list/skin/skin.ts
@@ -1,4 +1,4 @@
-import { List } from "../../List";
+import { List } from "../List";
 
 export class SkinList extends List {
     constructor() {
