+++ lib/list/title/title.ts
@@ -1,4 +1,4 @@
-import { List } from "../../List";
+import { List } from "../List";
 
 export class TitleList extends List {
     constructor() {
